Kernel/IPC: Use 51.14 microseconds for the SendSyncRequest delay approximation.
As measured by the time it takes for svcSendSyncRequest to return in an o3DS running firmware 11.6 using a homebrew that registers a dummy service that just responds after a sync request.

See https://gist.github.com/Subv/db8c299f5cf37704e360fbd1dd8eb895 for the source code of the test.