HLE/APT: Allow launching native applets instead of HLE ones. (#20)
* Logging: Add a Service_NS log class for the NS service.

* HLE/NS: Added a function to launch titles installed to the virtual NAND/SD card.

It uses AM::GetTitleContentPath to retrieve the path of the program to launch.

* HLE/APT: Added a function to retrieve the titleid of an applet based on the current system region.

The table was taken from the real APT service, but is incomplete due to the sheer amount of data it contains. There's 29 applets with 7 possible titleids. This table should be filled as needed.

* HLE/APT: Try to launch a native applet in PrepareToStartLibraryApplet and PreloadLibraryApplet before falling back to HLE applets.

With this commit, you can run native applets if they are in the correct folder of your virtual NAND.

Trying to exit the applet will currently cause an invalid read loop due to svcExitProcess not being implemented.

* NS/Applets: Added the MiiSelector applet to the list of native applet ids.