Kernel/IPC: Add a small delay after each SyncRequest to prevent thread starvation. (#61)
* Kernel/IPC: Add a small delay after each SyncRequest to prevent thread starvation.

In a real 3DS, threads that call svcSyncRequest are put to sleep until the server responds via svcReplyAndReceive. Our HLE services don't implement this mechanism and are effectively immediate from the 3DS's point of view. This commit makes it so that we at least simulate the IPC delay.

Specific HLE handlers might need to put their callers to sleep for a longer period of time to simulate IO timings. This is their responsibility but doing so is currently not implemented.

See https://gist.github.com/ds84182/4a7690c5376e045cab9129ca4185bbeb for a test that was not passing before this commit.

* Kernel/IPC: Use 51.14 microseconds for the SendSyncRequest delay approximation.

As measured by the time it takes for svcSendSyncRequest to return in an o3DS running firmware 11.6 using a homebrew that registers a dummy service that just responds after a sync request.

See https://gist.github.com/Subv/db8c299f5cf37704e360fbd1dd8eb895 for the source code of the test.